Reapply "Move imported token remove logic to the confirmation modal"
This reverts commit 2474637647b868e6686ed70680c6c122222fe258.
